requirements.txt
aircv==1.4.6
airtest==1.1.11
airtest-selenium==1.0.3
allure-pytest==2.8.6
allure-python-commons==2.8.6
Appium-Python-Client==1.0.1
asgiref==3.2.10
atomicwrites==1.4.0
attrs==21.2.0
cached-property==1.5.2
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
colorama==0.4.4
common==0.1.2
comtypes==1.1.10
ddt==1.4.2
decorator==5.0.7
Deprecated==1.2.12
Django==3.0.7
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.0.1
facebook-wda==1.4.0
fanstatic==1.1
ffmpy==0.2.3
gitdb2==2.0.6
GitPython==2.1.15
grpcio==1.32.0
hrpc==1.0.8
http-client==0.1.22
httplib2==0.18.1
idna==2.7
iniconfig==1.1.1
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
js==1.0
jsonpath==0.82
jsonschema==2.6.0
loguru==0.5.3
lxml==4.5.1
MarkupSafe==1.1.1
more-itertools==8.11.0
mss==4.0.3
Naked==0.1.31
nose==1.3.7
numpy==1.19.0
opencv-contrib-python==4.5.1.48
opencv-python==4.2.0.34
openpyxl==3.0.5
packaging==21.0
parameterized==0.8.1
Pillow==8.2.0
pluggy==0.13.1
poco==0.98.1
pocoui==1.0.82
protobuf==3.13.0
py==1.10.0
pyaml==18.11.0
pyaxmlparser==0.3.24
pycryptodome==3.9.8
pycurl==7.43.0.5
pyDes==2.0.1
PyExecJS==1.5.1
pygame==1.9.6
PyGithub==1.43.5
PyJWT==2.1.0
PyMySQL==0.9.3
pynput==1.7.3
pyparsing==2.4.7
pypiwin32==223
pytest==5.4.3
pytest-base-url==1.4.2
pytest-html==2.1.1
pytest-metadata==1.11.0
python-dateutil==2.8.1
python-gitlab==1.8.0
pytz==2020.1
pywin32==300
pywinauto==0.6.3
PyYAML==5.4.1
requests==2.20.1
requests-toolbelt==0.9.1
retry==0.9.2
selenium==3.141.0
shellescape==3.8.1
shutilwhich==1.1.0
six==1.15.0
smmap==4.0.0
smmap2==3.0.1
spark-parser==1.8.9
sqlparse==0.3.1
suds-jurko==0.6
svn==0.3.46
texttable==0.9.1
toml==0.10.2
uncompyle6==3.7.4
urllib3==1.24.3
vcode==0.0.2
wcwidth==0.2.5
WebOb==1.8.6
websocket-client==0.48.0
win32-setctime==1.0.3
wrapt==1.12.1
xdis==5.0.9
xlrd==1.2.0
xls==0.0.2
XlsxWriter==3.0.1
xlwt==1.3.0
